requirements.txt
catboost
Django
django-bootstrap3
fastapi
gunicorn
joblib
matplotlib
numpy
pandas
pexpect
Pillow
plotly
dj-database-url
whitenoise
scikit-learn
scipy
sklearn
twilio
google-api-python-client
django-heroku
google-api-core
google-auth
google-auth-httplib2
google-auth-oauthlib
google-pasta
googleapis-common-protos
